Move common video message handling logic into a base class Add thread safety and message handling to soft video processer
